+++ src/components/Product/ProductShort.tsx
@@ -2,37 +2,41 @@ import { FC } from 'react';
 import imgMonitor from '../../assets/img/monitor.png';
 import { Image } from 'react-bootstrap';
 import { BtnTrush } from '../BtnTrush/BtnTrush';
-import {
-  addHandleDeleteItem,
-  addItemToDelete,
-  handleDetailOrder,
-  handlePopUpOpen,
-} from '../../store/slices';
+import { addItemToDelete, handleDetailOrder, handlePopUpOpen } from '../../store/slices';
 import { IProduct } from '../../types/types';
 import { useAppDispatch } from '../../store/appDispatch';
 import './Product.scss';
 import { deleteProduct, fetchOrderList } from '../../store/api';
+import { useGetConfirmation } from '../../ConfirmationProvider';
 
 interface IProductProps {
   product: IProduct;
 }
 
 export const ProductShort: FC<IProductProps> = ({ product }) => {
+  const getConfirmation = useGetConfirmation();
   const dispatch = useAppDispatch();
 
-  const handleItemDelete = (): void => {
-    dispatch(deleteProduct(product.id));
-    dispatch(fetchOrderList());
-    dispatch(handleDetailOrder());
-    console.log('detail delete product');
-  };
+  // const handleItemDelete = (): void => {
+  //   dispatch(deleteProduct(product.id));
+  //   dispatch(fetchOrderList());
+  //   dispatch(handleDetailOrder());
+  //   console.log('detail delete product');
+  // };
 
-  const handleDeleteProduct = (): void => {
+  const handleDeleteProduct = async (): Promise<void> => {
     console.log('productShortTOdelete', product);
 
     dispatch(addItemToDelete(product));
     dispatch(handlePopUpOpen());
-    // dispatch(addHandleDeleteItem(handleItemDelete));
+    const confirmation = await getConfirmation({});
+
+    if (confirmation) {
+      dispatch(deleteProduct(product.id));
+      dispatch(fetchOrderList());
+      dispatch(handleDetailOrder());
+      console.log('OKDETAIL====>');
+    }
   };
 
   return (
